refactor: separate entity-block from display-entity component
to simplify components and remove duplications, removing separate display-entity-array
