+++ src/app/child-dev-project/attendance/dashboard-widgets/attendance-week-dashboard/attendance-week-dashboard.component.html
@@ -23,10 +23,10 @@
           (click)="goToChild(rowGroup[0].childId)"
           class="pointer"
         >
-          <app-display-entity
+          <app-entity-block
             [entityId]="rowGroup[0].childId"
             entityType="Child"
-          ></app-display-entity>
+          ></app-entity-block>
         </td>
       </ng-container>
