+++ src/app/child-dev-project/attendance/edit-attendance/edit-attendance.component.html
@@ -24,10 +24,10 @@
           </button>
         </td>
         <td>
-          <app-display-entity
+          <app-entity-block
             [entityId]="childId"
             entityType="Child"
-          ></app-display-entity>
+          ></app-entity-block>
         </td>
         <td>
           <app-attendance-status-select
@@ -66,10 +66,10 @@
     >
       <mat-card-content>
         <div class="attendance-item--header margin-bottom-regular">
-          <app-display-entity
+          <app-entity-block
             [entityId]="childId"
             entityType="Child"
-          ></app-display-entity>
+          ></app-entity-block>
 
           <button
             mat-icon-button
