+++ src/app/core/basic-datatypes/entity-array/display-entity/display-entity.component.scss
