+++ src/app/core/basic-datatypes/entity-array/edit-entity/edit-entity.stories.ts
@@ -17,7 +17,7 @@ const user2 = new User("2");
 user2.name = "other entity";
 
 export default {
-  title: "Core/Entities/Properties/entity-array/Entity Reference Array",
+  title: "Core/Entities/Properties/entity/EditEntity",
   component: FormComponent,
   decorators: [
     applicationConfig({
