--- src/app/core/basic-datatypes/entity/display-entity/display-entity.component.html
@@ -1,28 +1,0 @@
-<span
-  class="block-height truncate-text"
-  [ngClass]="{ clickable: !linkDisabled }"
-  (click)="showDetailsPage()"
->
-  <ng-container
-    *ngIf="entityBlockComponent && entityToDisplay"
-    [appDynamicComponent]="{
-      component: entityBlockComponent,
-      config: {
-        entity: entityToDisplay,
-        linkDisabled: linkDisabled,
-        tooltipDisabled: true
-      }
-    }"
-  >
-  </ng-container>
-
-  <ng-container *ngIf="!entityBlockComponent">
-    <app-fa-dynamic-icon
-      *ngIf="entityIcon"
-      [icon]="entityIcon"
-      class="margin-right-small"
-    ></app-fa-dynamic-icon>
-
-    {{ entityToDisplay?.toString() }}
-  </ng-container>
-</span>
