+++ src/app/core/basic-datatypes/entity/entity-block/entity-block.component.html
@@ -1,0 +1,27 @@
+<span
+  class="block-height truncate-text"
+  [ngClass]="{ clickable: !linkDisabled }"
+  (click)="showDetailsPage()"
+>
+  @if (entityBlockComponent) {
+    <ng-container
+      [appDynamicComponent]="{
+        component: entityBlockComponent,
+        config: {
+          entity: entityToDisplay,
+          linkDisabled: linkDisabled,
+          tooltipDisabled: true
+        }
+      }"
+    >
+    </ng-container>
+  } @else {
+    <app-fa-dynamic-icon
+      *ngIf="entityIcon"
+      [icon]="entityIcon"
+      class="margin-right-small"
+    ></app-fa-dynamic-icon>
+
+    {{ entityToDisplay?.toString() }}
+  }
+</span>
