+++ src/app/core/basic-datatypes/entity/entity-block/entity-block.component.scss
