+++ src/app/core/import/import-additional-actions/import-additional-actions.component.ts
@@ -16,7 +16,7 @@ import { MatListModule } from "@angular/material/list";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { NgForOf, NgIf } from "@angular/common";
-import { DisplayEntityComponent } from "../../basic-datatypes/entity/display-entity/display-entity.component";
+import { EntityBlockComponent } from "../../basic-datatypes/entity/entity-block/entity-block.component";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { BasicAutocompleteComponent } from "../../common-components/basic-autocomplete/basic-autocomplete.component";
 import { MatButtonModule } from "@angular/material/button";
@@ -35,7 +35,7 @@ import { MatButtonModule } from "@angular/material/button";
     MatTooltipModule,
     NgForOf,
     EntityTypeLabelPipe,
-    DisplayEntityComponent,
+    EntityBlockComponent,
     ReactiveFormsModule,
     MatFormFieldModule,
     BasicAutocompleteComponent,
