+++ src/app/core/import/import-history/import-history.component.html
@@ -18,11 +18,11 @@
 
     <div i18n>
       imported by
-      <app-display-entity
+      <app-entity-block
         [entityId]="item.user"
         entityType="User"
         [linkDisabled]="true"
-      ></app-display-entity>
+      ></app-entity-block>
       at {{ item.date | date: "medium" }}
     </div>
