+++ src/app/core/user/user-account/user-account.component.html
@@ -36,10 +36,10 @@
     <div *ngIf="entityId">
       <hr />
       <h2 i18n>Your profile:</h2>
-      <app-display-entity
+      <app-entity-block
         [entityId]="entityId"
         class="entity-box"
-      ></app-display-entity>
+      ></app-entity-block>
     </div>
   </mat-tab>
 </mat-tab-group>
