+++ src/app/features/dashboard-widgets/birthday-dashboard-widget/birthday-dashboard/birthday-dashboard.component.html
@@ -14,9 +14,9 @@
       </tr>
       <ng-container matColumnDef="entity">
         <td *matCellDef="let entity">
-          <app-display-entity
+          <app-entity-block
             [entityToDisplay]="entity.entity"
-          ></app-display-entity>
+          ></app-entity-block>
         </td>
       </ng-container>
       <ng-container matColumnDef="dateOfBirth">
